@@ -33,43 +33,19 @@ namespace Digipolis.Web.SampleApi.Controllers
         /// <remarks>
         /// This endpoint demostrates the use of the old <see cref="PagedResult{T}"/> and it's resulting json.
         /// </remarks>
-        /// <param name="queryOptions">Query options from uri</param>
         /// <returns>An array of value objects</returns>
         [HttpGet()]
         [ProducesResponseType(typeof(PagedResult<ValueDto>), 200)]
         [AllowAnonymous]
-        [Versions(Versions.V1, Versions.V2)]
         [Produces("application/hal+json")]
         public IActionResult Get([FromQuery]CriteriaDto criteria)
         {
             int total;
             var values = _valueLogic.GetAll(criteria, out total);
-            //var result = queryOptions.ToPagedResult(values, total, "kevin", new { test = 0 });
-            var result = criteria.ToPagedResult(values, total, "Get", "Values", new { test = 0 });
+            var result = criteria.ToPagedResult(values, total, ControllerContext.ActionDescriptor.ActionName, ControllerContext.ActionDescriptor.ControllerName);
             return Ok(result);
         }
-
-        /// <summary>
-        /// Get all values 
-        /// </summary>
-        /// <remarks>
-        /// This endpoint demostrates the use of the new <see cref="PagedResult{T, EmbeddedT}"/> and it's resulting json.
-        /// </remarks>
-        /// <param name="queryOptions">Query options from uri</param>
-        /// <returns>An array of value objects</returns>
-        [HttpGet("new")]
-        [ProducesResponseType(typeof(PagedResult<ValueDto, EmbeddedValueDto>), 200)]
-        [AllowAnonymous]
-        [Versions(Versions.V1, Versions.V2)]
-        [Produces("application/hal+json")]
-        public IActionResult GetNew([FromQuery]CriteriaDto criteria)
-        {
-            int total;
-            var values = _valueLogic.GetAll(criteria, out total);
-            //var result = queryOptions.ToPagedResult(values, total, "kevin", new { test = 0 });
-            var result = criteria.ToPagedResult<ValueDto, EmbeddedValueDto>(values, total);
-            return Ok(result);
-        }
+        
         /// <summary>
         /// Get a value by id
         /// </summary>
@@ -85,7 +61,7 @@ namespace Digipolis.Web.SampleApi.Controllers
             var value = _valueLogic.GetById(id);
             return Ok(value);
         }
-
+        
         /// <summary>
         /// Add a new value
         /// </summary>
@@ -95,13 +71,12 @@ namespace Digipolis.Web.SampleApi.Controllers
         [ValidateModelState]
         [ProducesResponseType(typeof(ValueDto), 201)]
         [AllowAnonymous]
-        [Versions(Versions.V1, Versions.V2)]
         public IActionResult Post([FromBody, Required] ValueDto value)
         {
             value = _valueLogic.Add(value);
             return CreatedAtAction("Get", new { id = value.Id }, value);
         }
-
+        
         /// <summary>
         /// Update an existing value object
         /// </summary>
@@ -110,34 +85,31 @@ namespace Digipolis.Web.SampleApi.Controllers
         /// <returns></returns>
         [HttpPut("{id}")]
         [ValidateModelState]
-        [Versions(Versions.V1, Versions.V2)]
         [ExcludeSwaggerResonse((int)HttpStatusCode.NotFound)]
         public IActionResult Put(int id, [FromBody, Required] ValueDto value)
         {
             value = _valueLogic.Update(id, value);
             return Ok(value);
         }
-
+        
         /// <summary>
         /// Delete a value by it's Id
         /// </summary>
         /// <param name="id">The value's Id</param>
         /// <returns></returns>
         [HttpDelete("{id}")]
-        [Versions(Versions.V2)]
         public IActionResult Delete(int id)
         {
             _valueLogic.Delete(id);
             return NoContent();
         }
-
+        
         /// <summary>
-        /// Thorws an exception
+        /// Throws an exception
         /// </summary>
         /// <returns></returns>
         [HttpGet("exception")]
         [AllowAnonymous]
-        [Versions(Versions.V1, Versions.V2)]
         public IActionResult ThrowException()
         {
             throw new NotFoundException();
