@@ -1,20 +1,12 @@
 ﻿using Digipolis.Web.Api;
-using Digipolis.Web.Api.Conventions;
-using Digipolis.Web.Api.JsonConverters;
 using Digipolis.Web.Api.Tools;
-using Digipolis.Web.Modelbinders;
 using Microsoft.AspNetCore.Http;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Formatters;
-using Microsoft.AspNetCore.Mvc.Infrastructure;
-using Microsoft.AspNetCore.Mvc.Routing;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.DependencyInjection.Extensions;
-using Swashbuckle.AspNetCore.SwaggerGen;
 using System;
-using System.Linq;
-using System.Reflection;
+using Microsoft.AspNetCore.Mvc.Infrastructure;
+using Microsoft.AspNetCore.Mvc.Routing;
 
 namespace Digipolis.Web
 {
@@ -28,11 +20,11 @@ namespace Digipolis.Web
 
             builder.Services.TryAddSingleton<IHttpContextAccessor, HttpContextAccessor>();
             builder.Services.TryAddSingleton<IActionContextAccessor, ActionContextAccessor>();
-
-            builder.Services.AddScoped<IUrlHelper>(x =>
-            {
-                var actionContext = x.GetService<IActionContextAccessor>().ActionContext;
-                return new UrlHelper(actionContext);
+            
+            builder.Services.AddScoped(sp => {
+                var ac = sp.GetRequiredService<IActionContextAccessor>().ActionContext;
+                var f = sp.GetRequiredService<IUrlHelperFactory>();
+                return f.GetUrlHelper(ac);
             });
 
             builder.Services.AddScoped<ILinkProvider, LinkProvider>();
@@ -49,52 +41,11 @@ namespace Digipolis.Web
             }
             if (build != null)
             {
-                builder.Services.Configure<ApiExtensionOptions>(build);
+                builder.Services.Configure(build);
                 build(apiOptions);
             }
 
             #endregion
-
-            #region Configuration from options
-
-            if (!apiOptions.DisableVersioning)
-            {
-                builder.AddMvcOptions(options =>
-                {
-                    options.Conventions.Insert(0, new RouteConvention(new RouteAttribute("{apiVersion}")));
-                });
-
-                builder.Services.ConfigureSwaggerGen(options =>
-                {
-                    options.DocInclusionPredicate((version, apiDescription) =>
-                    {
-                        if (!apiDescription.TryGetMethodInfo(out MethodInfo methodInfo)) return false;
-
-                        var allowedVersions = methodInfo.GetCustomAttributes(true).OfType<VersionsAttribute>().FirstOrDefault();
-                        return (allowedVersions != null && allowedVersions.AcceptedVersions.Contains(version));
-                    });
-                });
-            }
-
-
-            #endregion
-
-            builder.AddMvcOptions(options =>
-            {
-                options.Filters.Insert(0, new ConsumesAttribute("application/json"));
-                options.Filters.Insert(1, new ProducesAttribute("application/json"));
-
-                options.ModelBinderProviders.Insert(0, new CommaDelimitedArrayModelBinderProvider());
-
-                JsonOutputFormatter jsonFormatter = options.OutputFormatters.OfType<JsonOutputFormatter>().FirstOrDefault();
-
-                jsonFormatter?.SupportedMediaTypes.Add("application/hal+json");
-            });
-
-            builder.AddJsonOptions(x =>
-            {
-                x.SerializerSettings.Initialize();
-            });
 
             return builder;
         }
