@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 using Microsoft.AspNetCore.Http;
+using Microsoft.OpenApi.Models;
 using Swashbuckle.AspNetCore.SwaggerGen;
 using Swashbuckle.AspNetCore.Swagger;
 
@@ -10,7 +11,7 @@ namespace Digipolis.Web.Swagger
 {
     public class AddFileUploadParams : IOperationFilter
     {
-        public void Apply(Operation operation, OperationFilterContext context)
+        public void Apply(OpenApiOperation operation, OperationFilterContext context)
         {
             if (operation.Parameters == null)
                 return;
@@ -31,33 +32,35 @@ namespace Digipolis.Web.Swagger
             if (!allFileParamNames.Any())
                 return;
 
-            var paramsToRemove = new List<IParameter>();
-            foreach (var param in operation.Parameters)
-            {
-                paramsToRemove.AddRange(from fileParamName in allFileParamNames
-                    where param.Name.StartsWith(fileParamName + ".")
-                    select param);
-            }
-            paramsToRemove.ForEach(x => operation.Parameters.Remove(x));
-            foreach (var paramName in allFileParamNames)
-            {
-                var fileParam = new NonBodyParameter
-                {
-                    Type = "file",
-                    Name = paramName,
-                    In = "formData",
-                    Required = true
-                };
-                operation.Parameters.Add(fileParam);
-            }
-            foreach (
-                IParameter param in
-                operation.Parameters.Where(x => x.In.Equals("form", StringComparison.CurrentCultureIgnoreCase)))
-            {
-                param.In = "formData";
-            }
-
-            operation.Consumes = new List<string>() {"multipart/form-data"};
+            // var paramsToRemove = new List<OpenApiParameter>();
+            // foreach (var param in operation.Parameters)
+            // {
+            //     paramsToRemove.AddRange(from fileParamName in allFileParamNames
+            //         where param.Name.StartsWith(fileParamName + ".")
+            //         select param);
+            // }
+            //
+            // paramsToRemove.ForEach(x => operation.Parameters.Remove(x));
+            // foreach (var paramName in allFileParamNames)
+            // {
+            //     var fileParam = new OpenApiParameter()
+            //     {
+            //         Type = "file",
+            //         Name = paramName,
+            //         In = "formData",
+            //         Required = true
+            //     };
+            //     operation.Parameters.Add(fileParam);
+            // }
+            //
+            // foreach (
+            //     OpenApiParameter param in
+            //     operation.Parameters.Where(x => x.In.Equals("form", StringComparison.CurrentCultureIgnoreCase)))
+            // {
+            //     param.In = "formData";
+            // }
+            //
+            // // operation.Consumes = new List<string>() {"multipart/form-data"};
         }
     }
 }
